// src/generated/api/admin/models/adminUserDetailResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aller Meal API
 * Aller Meal Release 1 public, member, and admin API contract.
 * OpenAPI spec version: 1.0.0
 */
import type { AdminUserAvailableActionsResponse } from "./adminUserAvailableActionsResponse";
import type { AdminUserDetailResponseEmailVerificationStatus } from "./adminUserDetailResponseEmailVerificationStatus";
import type { AdminUserDetailResponseRole } from "./adminUserDetailResponseRole";
import type { AdminUserDetailResponseStatus } from "./adminUserDetailResponseStatus";

export interface AdminUserDetailResponse {
  userId: string;
  email: string;
  role: AdminUserDetailResponseRole;
  status: AdminUserDetailResponseStatus;
  emailVerificationStatus: AdminUserDetailResponseEmailVerificationStatus;
  createdAt: string;
  /** @nullable */
  withdrawalDueAt?: string | null;
  version: number;
  availableActions: AdminUserAvailableActionsResponse;
}